move full block write to complete_block
