+++ blockchain/miniblocks_consensus.go
@@ -206,6 +206,8 @@ func (chain *Blockchain) InsertMiniBlock(mbl block.MiniBlock) (err error, result
 		chain.RPC_NotifyNewMiniBlock.L.Unlock()
 
 		chain.flip_top()
+	} else {
+		fmt.Printf("error inserting miniblock to chain %s\n", err)
 	}
 	return err, result
 }
